Restructure project; Clean up code
