+++ Data/Models/ImpersonationRequest.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace Skrawl.API.Data.Models
+{
+    public class ImpersonationRequest
+    {
+        [JsonPropertyName("email")]
+        public string Email { get; set; }
+    }
+}
