+++ Data/Models/JwtAuthResult.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Skrawl.API.Infrastructure
+namespace Skrawl.API.Data.Models
 {
     public class JwtAuthResult
     {
