+++ Data/Models/JwtTokenConfig.cs
@@ -1,4 +1,4 @@
-namespace Skrawl.API.Infrastructure
+namespace Skrawl.API.Data.Models
 {
     public class JwtTokenConfig
     {
