+++ Data/Models/LoginRequest.cs
@@ -1,0 +1,16 @@
+using System.Text.Json.Serialization;
+using System.ComponentModel.DataAnnotations;
+
+namespace Skrawl.API.Data.Models
+{
+    public class LoginRequest
+    {
+        [Required]
+        [JsonPropertyName("email")]
+        public string Email { get; set; }
+
+        [Required]
+        [JsonPropertyName("password")]
+        public string Password { get; set; }
+    }
+}
