+++ Data/Models/LoginResult.cs
@@ -1,0 +1,22 @@
+using System.Text.Json.Serialization;
+
+namespace Skrawl.API.Data.Models
+{
+    public class LoginResult
+    {
+        [JsonPropertyName("email")]
+        public string Email { get; set; }
+
+        [JsonPropertyName("role")]
+        public string Role { get; set; }
+
+        [JsonPropertyName("originalEmail")]
+        public string OriginalEmail { get; set; }
+
+        [JsonPropertyName("accessToken")]
+        public string AccessToken { get; set; }
+
+        [JsonPropertyName("refreshToken")]
+        public string RefreshToken { get; set; }
+    }
+}
