+++ Data/Models/RefreshTokenRequest.cs
@@ -1,0 +1,10 @@
+using System.Text.Json.Serialization;
+
+namespace Skrawl.API.Data.Models
+{
+    public class RefreshTokenRequest
+    {
+        [JsonPropertyName("refreshToken")]
+        public string RefreshToken { get; set; }
+    }
+}
