+++ Infrastructure/JwtAuthManager.cs
@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
 using Microsoft.IdentityModel.Tokens;
+using Skrawl.API.Data.Models;
 
 namespace Skrawl.API.Infrastructure
 {
